backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
langchain==0.0.346
langchain-google-genai==0.0.1
google-generativeai==0.3.2
python-multipart==0.0.6
pydantic==1.10.12
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0